+++ recfm-java-lib/src/main/java/io/github/epi155/recfm/java/FixBasic.java
@@ -10,9 +10,8 @@ public interface FixBasic extends Validable {
      */
     String encode();
 
-//    /**
-//     * Record length
-//     * @return record length
-//     */
-//    int length();
+    /**
+     * initializes all fields to their default
+     */
+    void initialize();
 }